Fix 64-bit comparison
A 64-bit SIZE_MAX value cast to off64_t is always negative, causing valid
files to be rejected in 64-bit mode.

+++ media/libstagefright/SampleTable.cpp
@@ -278,9 +278,9 @@ status_t SampleTable::setSampleToChunkParams(
         return OK;
     }
 
-    if ((off64_t)(kMaxOffset - 8 -
+    if ((kMaxOffset - 8 -
             ((mNumSampleToChunkOffsets - 1) * sizeof(SampleToChunkEntry)))
-            < mSampleToChunkOffset) {
+            < (size_t)mSampleToChunkOffset) {
         return ERROR_MALFORMED;
     }
